Build System: new colorizer.py to color GCH
So I found a bug with colorizer.py when I first got it. I fixed
it somewhat cheaply at the time, but e-mailed the guy who wrote
the program telling him about the problem. As a result, he fixed
the bug and added a feature to be able to color any Scons
builder whatsoever. Now GCH has coloring for free... Bonus!